[MOVE/#55] 뒤로가기 종료 로직 이동
